+++ app/src/main/java/com/poupa/vinylmusicplayer/util/MusicUtil.java
@@ -209,7 +209,7 @@ public class MusicUtil {
         return trackNumberToFix % 1000;
     }
 
-    public static void insertAlbumArt(@NonNull Context context, int albumId, String path) {
+    public static void insertAlbumArt(@NonNull Context context, long albumId, String path) {
         ContentResolver contentResolver = context.getContentResolver();
 
         Uri artworkUri = Uri.parse("content://media/external/audio/albumart");
@@ -222,7 +222,7 @@ public class MusicUtil {
         contentResolver.insert(artworkUri, values);
     }
 
-    public static void deleteAlbumArt(@NonNull Context context, int albumId) {
+    public static void deleteAlbumArt(@NonNull Context context, long albumId) {
         ContentResolver contentResolver = context.getContentResolver();
         Uri localUri = Uri.parse("content://media/external/audio/albumart");
         contentResolver.delete(ContentUris.withAppendedId(localUri, albumId), null, null);
@@ -287,7 +287,7 @@ public class MusicUtil {
                     // as from the album art cache
                     cursor.moveToFirst();
                     while (!cursor.isAfterLast()) {
-                        final int id = cursor.getInt(0);
+                        final long id = cursor.getLong(0);
                         final Song song = SongLoader.getSong(activity, id);
                         MusicPlayerRemote.removeFromQueue(song);
                         cursor.moveToNext();
@@ -380,7 +380,7 @@ public class MusicUtil {
         return String.valueOf(musicMediaTitle.charAt(0)).toUpperCase();
     }
 
-    public static int indexOfSongInList(List<Song> songs, int songId) {
+    public static int indexOfSongInList(List<Song> songs, long songId) {
         for (int i = 0; i < songs.size(); i++) {
             if (songs.get(i).id == songId) {
                 return i;
